refactor(mcp): split OntologyServerCapabilitiesProvider out of OntologyToolDiscovery (M6)
OntologyToolDiscovery had two responsibilities — discovering tools and
exposing server initialization metadata. Splits server-capability exposure
into a dedicated OntologyServerCapabilitiesProvider so each type owns one
concern. The provider validates its OntologyGraph dependency eagerly via
ArgumentNullException.ThrowIfNull, surfacing composition-time bugs.

Tests moved from OntologyToolDiscoveryAnnotationTests to a new
OntologyServerCapabilitiesProviderTests file with the existing
GetServerCapabilities_ReturnsCurrentGraphVersion plus a new
OntologyServerCapabilitiesProvider_NullGraph_Throws guard.

Design doc §4.8 and §9 + plan Task B9 + acceptance criteria updated to
reflect the SRP split.
